wip: SCOR-89 okay serverless-appsync-plugin does support
appsync_js runtime for resolvers, per Javascript Vs Vtl here
https://github.com/sid88in/serverless-appsync-plugin/blob/master/doc/resolvers.md
but this config is all bootsy and from the past.  first checkpointing where I was but need to go back to how it was before with all VTL and just try to move the config to custom like the framework expects now.