build: SCOR-82 remove the nonce
because it is a lot of effort for very little reward. we have to deal with an additional api_key authorized endpoint, with its expiration of one year. We should probably do a recaptcha on the tablet since it has to make an api_key authorized call. The api has the additional endpoints with their maintenance cost.  All to input four or five rather than sixteen characters?  Silly.